use separate editor.cfg configuration file to specify editable file(s)
For every user (as authenticated via HTTP Auth) the config file can specify
one file that the user may edit.

This implementation uses a hand-made config file parser (ugh...) as there
appears to be no usable parser available in a stock Perl installation, and
the script is supposed to work also with simple CGI hosting services.